+++ live_reload.py
@@ -24,9 +24,9 @@ kpLiveReloadPause = bytearray(b'\xc5\xd9\x00\x00\x00\x00\x00\x00\x22')
 kpLiveReloadLevel = bytearray(b'\x4c\xc8\x00\x00\x00\x00\x00\x00\x23')
 kpLiveReloadStart = bytearray(b'\xf3\xbc\x00\x00\x00\x00\x00\x00\x24')
 
-# This is the expected version of Avara. Update this whenever you change the
-# above CRC bytes!
-expectedAvaraVersion = '36a5afb0'
+# This is the expected version of Avara. Update this whenever you update the
+# Avara editing-tools branch.
+expectedAvaraVersion = '92ed28d5'
 
 ip = '127.0.0.1'
 port_avara = 19567